pyproject.toml
[tool.poetry]
name = "transformer"
version = "0.1.0"
description = ""
authors = ["wideraHannes <jowid100@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.5.0"
torchvision = "^0.20.0"
torchaudio = "^2.5.0"
pytest = "^8.3.3"
tqdm = "^4.67.0"
tokenizers = "^0.20.3"
datasets = "^3.1.0"
transformers = "^4.46.2"
pyyaml = "^6.0.2"
tensorboard = "^2.18.0"
wandb = "^0.19.2"
nltk = "^3.9.1"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
rouge = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
